material.cpp

+++ src/material.h
@@ -59,7 +59,7 @@ struct Entry {
   const EndgameBase<Value>* evaluationFunction;
   const EndgameBase<ScaleFactor>* scalingFunction[COLOR_NB]; // Could be one for each
                                                              // side (e.g. KPKP, KBPsK)
-  int16_t value;
+  int32_t value;
   uint8_t factor[COLOR_NB];
   Phase gamePhase;
 };
